/* Composition is used in place of inheritance when a class is made up of classes that are not subclasses to that superclass.
in this example we build a computer out of a motherboard, a monitor, and a case. Both the monitor and case have their own
subcompisitions in Resolution and Dimensions. The Resolution is a component of the Monitor just as the Monitor is a component
of the PC. In our Monitor deceleration we also declare the resolution in our constructor. This is useful when we do not need
to access that object anywhere but within our composed class.
*
*
* */